requirements.txt
aiogram==3.17.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
lxml
PyYAML==6.0.1